refactor: consolidate find_next_run_dir function into io module and update usage in predict
